+++ src/de.secucard.connect/client-browser-environment.js
@@ -57,7 +57,7 @@ export const ClientBrowserEnvironment = {
         General.StoreGroupService,
         General.StoreService,
         General.TransactionService,
-        
+
         Loyalty.ActionService,
         Loyalty.ActionProfileService,
         Loyalty.ActionMessageService,
@@ -81,9 +81,11 @@ export const ClientBrowserEnvironment = {
         Payment.ContainerService,
         Payment.ContractService,
         Payment.CustomerService,
+        Payment.EterminalTransactionService,
         Payment.InvoiceService,
         Payment.PayoutService,
         Payment.SecupayDebitService,
+        Payment.SecupayPayoutService,
         Payment.SecupayPrepayService,
         Payment.TransactionService,
         Payment.TransactionHistoriesService,
@@ -94,12 +96,12 @@ export const ClientBrowserEnvironment = {
         Prepaid.ReportService,
         Prepaid.SaleService,
         Prepaid.StockService,
-        
+
         Services.IdentCaseService,
         Services.IdentContractService,
         Services.IdentRequestService,
         Services.IdentResultService,
-        
+
         Smart.CheckinService,
         Smart.ConfigurationService,
         Smart.DeviceService,
@@ -181,12 +183,14 @@ export const ServiceMap = {
         Containers: Payment.ContainerService.Uid,
         Contracts: Payment.ContractService.Uid,
         Customers: Payment.CustomerService.Uid,
+        EterminalTransactions: Payment.EterminalTransactionService.Uid,
         Invoices: Payment.InvoiceService.Uid,
         Payouts: Payment.PayoutService.Uid,
         SecupayDebits: Payment.SecupayDebitService.Uid,
+        SecupayPayouts: Payment.SecupayPayoutService.Uid,
         SecupayPrepays: Payment.SecupayPrepayService.Uid,
         Transactions: Payment.TransactionService.Uid,
-        TransactionHistories: Payment.TransactionHistoriesService.Uid
+        TransactionHistories: Payment.TransactionHistoriesService.Uid,
     },
     Prepaid:{
         Contracts: Prepaid.ContractService.Uid,