+++ src/de.secucard.connect/product/payment/eterminal-transaction-service.js
@@ -9,6 +9,22 @@
  See the License for the specific language governing permissions and
  limitations under the License.
  */
-export const Version = {
-    "name": "0.6.6"
-};
+import {ProductService} from '../product-service'
+
+export class EterminalTransactionService extends ProductService {
+
+    constructor() {
+        super()
+    }
+
+    getEndpoint() {
+        return ['payment', 'eterminaltransactions'];
+    }
+
+    getEventTargets() {
+        return [];
+    }
+
+}
+
+EterminalTransactionService.Uid = (['payment', 'eterminaltransactions']).join('.');