+++ src/de.secucard.connect/product/payment/secupay-payout-service.js
@@ -9,6 +9,25 @@
  See the License for the specific language governing permissions and
  limitations under the License.
  */
-export const Version = {
-    "name": "0.6.6"
-};
+import {ProductService} from '../product-service'
+
+export class SecupayPayoutService extends ProductService {
+
+    constructor() {
+        super()
+    }
+
+    getEndpoint() {
+        return ['payment', 'secupaypayout'];
+    }
+
+    getEventTargets() {
+        return [];
+    }
+
+    payoutWithoutCustomer(data) {
+        return this.execute('me', 'PayoutWithoutCustomer', null, data);
+    }
+}
+
+SecupayPayoutService.Uid = (['payment', 'secupaypayout']).join('.');